removed the cards behind the logo and intro text, made the darkmode go to the bottom of the page, moved the landing pages's project card into it's own folder
